+++ moabb/datasets/huebner_llp.py
@@ -117,15 +117,6 @@ class Huebner2017(_BaseVisualMatrixSpellerDataset):
     """Learning from label proportions for a visual matrix speller (ERP)
     dataset from Hübner et al 2017 [1]_.
 
-    .. admonition:: Dataset summary
-
-
-        ===========  =======  =======  =================  ===============  ===============  ===========
-        Name           #Subj    #Chan  #Trials / class    Trials length    Sampling rate      #Sessions
-        ===========  =======  =======  =================  ===============  ===============  ===========
-        Huebner2017       13       31  364 NT / 112 T     0.9s             1000Hz                     3
-        ===========  =======  =======  =================  ===============  ===============  ===========
-
     **Dataset description**
 
     The subjects were asked to spell the sentence: “Franzy jagt im komplett verwahrlosten Taxi quer durch Freiburg”.
@@ -183,15 +174,6 @@ class Huebner2017(_BaseVisualMatrixSpellerDataset):
 class Huebner2018(_BaseVisualMatrixSpellerDataset):
     """Mixture of LLP and EM for a visual matrix speller (ERP) dataset from
     Hübner et al 2018 [1]_.
-
-    .. admonition:: Dataset summary
-
-
-        ===========  =======  =======  =================  ===============  ===============  ===========
-        Name           #Subj    #Chan  #Trials / class    Trials length    Sampling rate      #Sessions
-        ===========  =======  =======  =================  ===============  ===============  ===========
-        Huebner2018       12       31  364 NT / 112 T     0.9s             1000Hz                     3
-        ===========  =======  =======  =================  ===============  ===============  ===========
 
     **Dataset description**